refactor dnd helpers
